requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.5.2               # via kombu
backcall==0.1.0           # via ipython
billiard==3.6.1.0         # via celery
brotlipy==0.7.0
celery[redis]==4.3.0
certifi==2019.11.28       # via requests, sentry-sdk
cffi==1.13.2              # via brotlipy
chardet==3.0.4            # via requests
decorator==4.4.1          # via ipython, traitlets
defusedxml==0.6.0         # via odfpy, python3-openid, social-auth-core
deprecated==1.2.7         # via pygithub
diff-match-patch==20181111  # via django-import-export
dj-database-url==0.5.0
django-cors-headers==3.2.0
django-import-export==1.2.0
django-js-reverse==0.9.1
django-log-request-id==1.4.0
django-model-utils==3.2.0
django-webpack-loader==0.6.0
django==2.2.7
djangorestframework==3.10.3
et-xmlfile==1.0.1         # via openpyxl
gunicorn==20.0.4
idna==2.8                 # via requests
importlib-metadata==0.23  # via kombu
ipython-genutils==0.2.0   # via traitlets
ipython==7.10.0
jdcal==1.4.1              # via openpyxl
jedi==0.15.1              # via ipython
kombu==4.6.6              # via celery
markuppy==1.14            # via tablib
more-itertools==7.2.0     # via zipp
oauthlib==3.1.0           # via requests-oauthlib, social-auth-core
odfpy==1.4.0              # via tablib
openpyxl==3.0.2           # via tablib
parso==0.5.1              # via jedi
pexpect==4.7.0            # via ipython
pickleshare==0.7.5        # via ipython
prompt-toolkit==3.0.1     # via ipython
psutil==5.6.7
psycopg2==2.8.4
ptyprocess==0.6.0         # via pexpect
pycparser==2.19           # via cffi
pygithub==1.44.1
pygments==2.5.2           # via ipython
pyjwt==1.7.1              # via pygithub, social-auth-core
python-decouple==3.3
python3-openid==3.1.0     # via social-auth-core
pytz==2019.3              # via celery, django
pyyaml==5.1.2             # via tablib
redis==3.3.11             # via celery
requests-oauthlib==1.3.0  # via social-auth-core
requests==2.22.0          # via pygithub, requests-oauthlib, social-auth-core
sentry-sdk==0.13.4
six==1.13.0               # via pygithub, social-auth-app-django, social-auth-core, traitlets
social-auth-app-django==3.1.0
social-auth-core==3.2.0   # via social-auth-app-django
sqlparse==0.3.0           # via django
tablib==0.14.0            # via django-import-export
traitlets==4.3.3          # via ipython
urllib3==1.25.7           # via requests, sentry-sdk
vine==1.3.0               # via amqp, celery
wcwidth==0.1.7            # via prompt-toolkit
whitenoise==4.1.4
wrapt==1.11.2             # via deprecated
xlrd==1.2.0               # via tablib
xlwt==1.3.0               # via tablib
zipp==0.6.0               # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools==42.0.1        # via gunicorn, ipython
